pyproject.toml
[tool.poetry]
name = "simple-api-flask"
version = "0.1.0"
description = "aula de Flask"
authors = ["adaltospjr <adaltospjr@icloud.com>"]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^1.1.2"
pylint = "^2.6.0"
python-dotenv = "^0.14.0"
psycopg2-binary = "^2.8.6"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
